Issue #3736 - move purchases to new area
Fixes #3736
These changes include:
Needed to change the dashboard to not contain any information about the purchases summary - this has now been moved to the new purchases summary page within the nav bar under `reporting and auditing`. Filtering has also been added to said page, and filter partials and the card have been added to a shared area for ease of access and consistency. Page tests added for new content.
