+++ app/views/dashboard/index.html.erb
@@ -313,26 +313,6 @@
 
         <%= render(
           "card",
-          id: "purchases",
-          gradient: "secondary",
-          title: "Purchases",
-          subtitle: @selected_date_range,
-          footer_options: { class: "text-center" },
-          footer: link_to("See more...", purchases_path),
-        ) do %>
-          <%= new_button_to new_purchase_path, {text: "New Purchase"} %>
-          <h3 class="text-center"><%= dollar_presentation(@purchases.sum(&:amount_spent_in_cents)) %>
-            spent
-            <%= @selected_date_range_label %>
-          </h3>
-          <div class="box-body">
-            <h4>Recent purchases</h4>
-            <%= render partial: "purchase", collection: @recent_purchases, as: :purchase %>
-          </div>
-        <% end %>
-
-        <%= render(
-          "card",
           gradient: "warning",
           title: "Itemized Donations",
           subtitle: @selected_date_range,
