+++ app/views/layouts/_lte_sidebar.html.erb
@@ -170,8 +170,8 @@
       </li>
     </ul>
   </li>
-  <li class="nav-item has-treeview <%= menu_open?(['audits', 'reports']) %>">
-    <a href="#" class="nav-link <%= active_class(['audits', 'reports']) %>">
+  <li class="nav-item has-treeview <%= menu_open?(['audits', 'reports', "purchases_summary"]) %>">
+    <a href="#" class="nav-link <%= active_class(['audits', 'reports', "purchases_summary"]) %>">
       <i class="nav-icon fa fa-tasks"></i>
       <p>Reporting & Auditing
         <i class="fas fa-angle-left right"></i>
@@ -193,6 +193,11 @@
       <li class="nav-item <%= active_class(['reports']) %>">
         <%= link_to(distributions_by_county_report_path(filters: { date_range: date_range_params }), class: "nav-link #{active_class(['reports'])}") do %>
           <i class="nav-icon fa fa-circle-o"></i> Distributions by County
+        <% end %>
+      </li>
+      <li class="nav-item <%= active_class(['purchases_summary']) %>">
+        <%= link_to(purchases_summary_path(organization_id: current_user.organization), class: "nav-link #{"active" if current_page?(purchases_summary_path)}") do %>
+          <i class="nav-icon fa fa-circle-o"></i> Purchases Summary
         <% end %>
       </li>
     </ul>
