+++ app/views/purchases_summary/index.html.erb
@@ -1,0 +1,20 @@
+        <%= render(
+          "shared/filtered_card",
+          id: "purchases",
+          gradient: "secondary",
+          title: "Purchases",
+          subtitle: @selected_date_range,
+          footer_options: { class: "text-center" },
+          footer: link_to("See more...", purchases_path),
+          filter_url: purchases_summary_path
+        ) do %>
+          <%= new_button_to new_purchase_path, {text: "New Purchase"} %>
+          <h3 class="text-center"><%= dollar_presentation(@purchases.sum(&:amount_spent_in_cents)) %>
+            spent
+            <%= @selected_date_range_label %>
+          </h3>
+          <div class="box-body">
+            <h4>Recent purchases</h4>
+            <%= render partial: "purchase", collection: @recent_purchases, as: :purchase %>
+          </div>
+        <% end %>
