+++ config/routes.rb
@@ -216,6 +216,7 @@ Rails.application.routes.draw do
     get "historical_trends/distributions", to: "historical_trends/distributions#index"
     get "historical_trends/purchases", to: "historical_trends/purchases#index"
     get "historical_trends/donations", to: "historical_trends/donations#index"
+    get "purchases_summary", to: "purchases_summary#index"
   end
 
   resources :attachments, only: %i(destroy)
